+++ main.tf
@@ -3,42 +3,6 @@
 # This module creates a GitHub repository with opinionated default settings.
 # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-# Set some opinionated default settings through var.defaults and locals
-locals {
-  homepage_url           = var.homepage_url == null ? lookup(var.defaults, "homepage_url", "") : var.homepage_url
-  private                = var.private == null ? lookup(var.defaults, "private", true) : var.private
-  private_visibility     = local.private ? "private" : "public"
-  visibility             = var.visibility == null ? lookup(var.defaults, "visibility", local.private_visibility) : var.visibility
-  has_issues             = var.has_issues == null ? lookup(var.defaults, "has_issues", false) : var.has_issues
-  has_projects           = var.has_projects == null ? lookup(var.defaults, "has_projects", false) : length(var.projects) > 0 ? true : var.has_projects
-  has_wiki               = var.has_wiki == null ? lookup(var.defaults, "has_wiki", false) : var.has_wiki
-  allow_merge_commit     = var.allow_merge_commit == null ? lookup(var.defaults, "allow_merge_commit", true) : var.allow_merge_commit
-  allow_rebase_merge     = var.allow_rebase_merge == null ? lookup(var.defaults, "allow_rebase_merge", false) : var.allow_rebase_merge
-  allow_squash_merge     = var.allow_squash_merge == null ? lookup(var.defaults, "allow_squash_merge", false) : var.allow_squash_merge
-  allow_auto_merge       = var.allow_auto_merge == null ? lookup(var.defaults, "allow_auto_merge", false) : var.allow_auto_merge
-  delete_branch_on_merge = var.delete_branch_on_merge == null ? lookup(var.defaults, "delete_branch_on_merge", true) : var.delete_branch_on_merge
-  is_template            = var.is_template == null ? lookup(var.defaults, "is_template", false) : var.is_template
-  has_downloads          = var.has_downloads == null ? lookup(var.defaults, "has_downloads", false) : var.has_downloads
-  auto_init              = var.auto_init == null ? lookup(var.defaults, "auto_init", true) : var.auto_init
-  gitignore_template     = var.gitignore_template == null ? lookup(var.defaults, "gitignore_template", "") : var.gitignore_template
-  license_template       = var.license_template == null ? lookup(var.defaults, "license_template", "") : var.license_template
-  default_branch         = var.default_branch == null ? lookup(var.defaults, "default_branch", null) : var.default_branch
-  standard_topics        = var.topics == null ? lookup(var.defaults, "topics", []) : var.topics
-  topics                 = concat(local.standard_topics, var.extra_topics)
-  template               = var.template == null ? [] : [var.template]
-  issue_labels_create    = var.issue_labels_create == null ? lookup(var.defaults, "issue_labels_create", local.issue_labels_create_computed) : var.issue_labels_create
-
-  issue_labels_create_computed = local.has_issues || length(var.issue_labels) > 0
-
-  # for readability
-  var_gh_labels = var.issue_labels_merge_with_github_labels
-  gh_labels     = local.var_gh_labels == null ? lookup(var.defaults, "issue_labels_merge_with_github_labels", true) : local.var_gh_labels
-
-  issue_labels_merge_with_github_labels = local.gh_labels
-  # Per default, GitHub activates vulnerability  alerts for public repositories and disables it for private repositories
-  vulnerability_alerts = var.vulnerability_alerts != null ? var.vulnerability_alerts : local.private ? false : true
-}
-
 locals {
   branch_protections_v3 = [
     for b in var.branch_protections_v3 : merge({
@@ -91,29 +55,28 @@ locals {
 resource "github_repository" "repository" {
   name                   = var.name
   description            = var.description
-  homepage_url           = local.homepage_url
-  visibility             = local.visibility
-  has_issues             = local.has_issues
-  has_projects           = local.has_projects
-  has_wiki               = local.has_wiki
-  allow_merge_commit     = local.allow_merge_commit
-  allow_rebase_merge     = local.allow_rebase_merge
-  allow_squash_merge     = local.allow_squash_merge
-  allow_auto_merge       = local.allow_auto_merge
-  delete_branch_on_merge = local.delete_branch_on_merge
-  is_template            = local.is_template
-  has_downloads          = local.has_downloads
-  auto_init              = local.auto_init
-  gitignore_template     = local.gitignore_template
-  license_template       = local.license_template
+  homepage_url           = var.homepage_url
+  visibility             = var.visibility
+  has_issues             = var.has_issues
+  has_projects           = length(var.projects) > 0 ? true : var.has_projects
+  has_wiki               = var.has_wiki
+  allow_merge_commit     = var.allow_merge_commit
+  allow_rebase_merge     = var.allow_rebase_merge
+  allow_squash_merge     = var.allow_squash_merge
+  allow_auto_merge       = var.allow_auto_merge
+  delete_branch_on_merge = var.delete_branch_on_merge
+  is_template            = var.is_template
+  has_downloads          = var.has_downloads
+  auto_init              = var.auto_init
+  gitignore_template     = var.gitignore_template
+  license_template       = var.license_template
   archived               = var.archived
-  topics                 = local.topics
-
-  archive_on_destroy   = var.archive_on_destroy
-  vulnerability_alerts = local.vulnerability_alerts
+  topics                 = concat(var.topics, var.extra_topics)
+  archive_on_destroy     = var.archive_on_destroy
+  vulnerability_alerts   = var.vulnerability_alerts
 
   dynamic "template" {
-    for_each = local.template
+    for_each = var.template == null ? [] : [var.template]
 
     content {
       owner      = template.value.owner
@@ -167,10 +130,10 @@ resource "github_branch" "branch" {
 # ---------------------------------------------------------------------------------------------------------------------
 
 resource "github_branch_default" "default" {
-  count = local.default_branch != null ? 1 : 0
+  count = var.default_branch != null ? 1 : 0
 
   repository = github_repository.repository.name
-  branch     = local.default_branch
+  branch     = var.default_branch
 
   depends_on = [github_branch.branch]
 }
@@ -296,7 +259,7 @@ locals {
   # all deployed repositories.
   # add labels if new labels in github are added by default.
   # this is the set of labels and colors as of 2020-02-02
-  github_default_issue_labels = local.issue_labels_merge_with_github_labels ? [
+  github_default_issue_labels = var.issue_labels_merge_with_github_labels ? [
     {
       name        = "bug"
       description = "Something isn't working"
@@ -354,7 +317,7 @@ locals {
 }
 
 resource "github_issue_label" "label" {
-  for_each = local.issue_labels_create ? local.issue_labels : {}
+  for_each = ((var.issue_labels_create == null ? false : var.issue_labels_create) || var.has_issues || length(var.issue_labels) > 0) ? local.issue_labels : {}
 
   repository  = github_repository.repository.name
   name        = each.value.name